+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## v0.3.32 - Combinations
-  * Commit: 2026-05-22 06:55 
+  * Commit: 2026-05-22 06:59 
   * Combinations
+  * RangeCombinations
 ## v0.3.31 - Multi-Column Order
   * Commit: 2026-05-18 09:56
   * Schema GetOrderedRows